Fix menu,DAO attContato,Exceptions

+++ src/main/java/dao/ContatoDAO.java
@@ -1,31 +1,83 @@
 package dao;
 
+import exceptions.CttNaoEncontradoException;
+import exceptions.DeletarException;
 import jakarta.persistence.EntityManager;
 import modelo.Contatos;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Scanner;
 
 public class ContatoDAO {
 
+
     private EntityManager em;
 
+    //Construtor
     public ContatoDAO(EntityManager em){
         this.em = em;
     }
 
+    //Metodos
     public void cadastrarContato(Contatos contatos){
         this.em.persist(contatos);
+        System.out.println("Contato salvo com sucesso \n");
     }
 
-    public Contatos atualizarContato(int id,String nome){
-        Contatos a = em.find(Contatos.class, id);
-        return a;
+    public void attContato(int id){
+        Scanner ler = new Scanner(System.in);
+        Contatos contatoAtualizado =  this.em.find(Contatos.class, id);
+
+        System.out.println("""
+                    1- Atualizar nome
+                    2- Atualizar numero
+                    3- Atualizar e-mail
+                    """);
+        int resp = ler.nextInt();
+        ler.nextLine();
+
+        if (resp == 1) {
+            System.out.println("Informe um nome");
+            String nome = ler.nextLine();
+
+            this.em.getTransaction().begin();
+            contatoAtualizado = this.em.merge(contatoAtualizado);
+            contatoAtualizado.setNome(nome);
+            this.em.getTransaction().commit();
+            this.em.close();
+            System.out.println("Nome atualizado: " + contatoAtualizado.getNome() + "\n" );
+        } else if (resp == 2) {
+            System.out.println("Informe o numero");
+            String numero = ler.nextLine();
+            this.em.getTransaction().begin();
+            contatoAtualizado = this.em.merge(contatoAtualizado);
+            contatoAtualizado.setTelefone(numero);
+            this.em.getTransaction().commit();
+            this.em.close();
+            System.out.println("Numero atualizado: " + contatoAtualizado.getTelefone() + "\n");
+        } else if (resp == 3) {
+            System.out.println("Informe o e-mail");
+            String email = ler.nextLine();
+            this.em.getTransaction().begin();
+            contatoAtualizado = this.em.merge(contatoAtualizado);
+            contatoAtualizado.setEmail(email);
+            this.em.getTransaction().commit();
+            this.em.close();
+            System.out.println("E-mail atualizado: " + contatoAtualizado.getEmail() + "\n");
+        }
     }
 
     public String buscarPorId(int id){
-      Contatos a = em.find(Contatos.class, id);
-      return a.toString();
+        Contatos a =this.em.find(Contatos.class, id);
+        if(a == null){
+            //Exception
+            CttNaoEncontradoException resp = new CttNaoEncontradoException("Contato não encontrado");
+            return resp.getMessage();
+        }else{
+            return a.toString();
+        }
+
     }
 
     public List<Contatos> todosContatos(){
@@ -42,6 +94,12 @@ public class ContatoDAO {
 
     public void deletarContato(int id){
         Contatos a = this.em.find(Contatos.class, id);
-        this.em.remove(a);
+        if (a == null){
+            DeletarException deletarException = new DeletarException("Contato não existe");
+            System.out.println(deletarException.getMessage() + '\n');
+        }else{
+            this.em.remove(a);
+        }
+
     }
 }
